requirements.txt
asgiref==3.7.2
chardet==5.2.0
Django==5.0.1
django-cors-headers==4.3.1
djangorestframework==3.14.0
pillow==10.2.0
pytz==2023.4
reportlab==4.0.9
sqlparse==0.4.4
tzdata==2023.4